Fix failing chained runs (based on runs)
The bug only happened when source-dir was used. The deploy like runs
would use the dir specified by source-dir as the output, causing
subsequent runs to fail because it's likely that the dir specified by
source-dir would not exist there.

The behaviour for collecting data should now be the same for deploys
as it is for builds

resolves #151