fix: (wercker/envvars#55)  Remove godotenv dep and load env file in order ourselves
godotcom was loading the environment file into a map and iterating  over that map to add the env key/val pairs back into the processes environment (thourhg os.setenv).

Iterating over a map is a randomized order in golang.

So if you have "NAME=Simon" and "GREETING=Hello $NAME" in the file then sometimes you'd get "GREETING=Hello " and sometimes it'd be "GREETING=Hello Simon" depending on the randomized order when iterating over the map.

So i've added a simplified file loading function to util.Environment