Keep the order when collecting pass-through vars
The GetPassthru and GetHiddenPassthru functions don't keep the order of
the Envs, that is because they range over the Environment.Map. The fix
is to range over the Environment.Order.
In addition to that we added a small refactor to put the logic of the
collection mechanism in a single function so we can avoid code
duplication.